+++ .github/workflows/build.yml
@@ -417,8 +417,8 @@ jobs:
         [ -n "$NGTCP2_BUILD_DIR" ] && cd "$NGTCP2_BUILD_DIR"
         "$NGTCP2_SOURCE_DIR"/ci/gen-certificate.sh
 
-        CLIENTS="client gtlsclient bsslclient wsslclient ptlsclient"
-        SERVERS="server gtlsserver bsslserver wsslserver ptlsserver"
+        CLIENTS="qtlsclient gtlsclient bsslclient wsslclient ptlsclient"
+        SERVERS="qtlsserver gtlsserver bsslserver wsslserver ptlsserver"
 
         for client in $CLIENTS; do
           for server in $SERVERS; do
