+++ CMakeLists.txt
@@ -36,9 +36,9 @@ set(LT_CURRENT  13)
 set(LT_REVISION 0)
 set(LT_AGE      0)
 
-set(CRYPTO_OPENSSL_LT_CURRENT 5)
-set(CRYPTO_OPENSSL_LT_REVISION 0)
-set(CRYPTO_OPENSSL_LT_AGE 0)
+set(CRYPTO_QUICTLS_LT_CURRENT 0)
+set(CRYPTO_QUICTLS_LT_REVISION 0)
+set(CRYPTO_QUICTLS_LT_AGE 0)
 
 set(CRYPTO_GNUTLS_LT_CURRENT 5)
 set(CRYPTO_GNUTLS_LT_REVISION 0)
@@ -54,10 +54,10 @@ include(Version)
 math(EXPR LT_SOVERSION "${LT_CURRENT} - ${LT_AGE}")
 set(LT_VERSION "${LT_SOVERSION}.${LT_AGE}.${LT_REVISION}")
 
-math(EXPR CRYPTO_OPENSSL_LT_SOVERSION
-  "${CRYPTO_OPENSSL_LT_CURRENT} - ${CRYPTO_OPENSSL_LT_AGE}")
-set(CRYPTO_OPENSSL_LT_VERSION
-  "${CRYPTO_OPENSSL_LT_SOVERSION}.${CRYPTO_OPENSSL_LT_AGE}.${CRYPTO_OPENSSL_LT_REVISION}")
+math(EXPR CRYPTO_QUICTLS_LT_SOVERSION
+  "${CRYPTO_QUICTLS_LT_CURRENT} - ${CRYPTO_QUICTLS_LT_AGE}")
+set(CRYPTO_QUICTLS_LT_VERSION
+  "${CRYPTO_QUICTLS_LT_SOVERSION}.${CRYPTO_QUICTLS_LT_AGE}.${CRYPTO_QUICTLS_LT_REVISION}")
 
 math(EXPR CRYPTO_GNUTLS_LT_SOVERSION
   "${CRYPTO_GNUTLS_LT_CURRENT} - ${CRYPTO_GNUTLS_LT_AGE}")
