+++ crypto/includes/Makefile.am
@@ -25,7 +25,7 @@ EXTRA_DIST = CMakeLists.txt
 nobase_include_HEADERS = ngtcp2/ngtcp2_crypto.h
 
 if HAVE_OPENSSL
-nobase_include_HEADERS += ngtcp2/ngtcp2_crypto_openssl.h
+nobase_include_HEADERS += ngtcp2/ngtcp2_crypto_quictls.h
 endif
 
 if HAVE_GNUTLS
