+++ crypto/quictls/CMakeLists.txt
@@ -23,12 +23,12 @@
 
 add_definitions(-DBUILDING_NGTCP2)
 
-set(ngtcp2_crypto_openssl_SOURCES
-  openssl.c
+set(ngtcp2_crypto_quictls_SOURCES
+  quictls.c
   ../shared.c
 )
 
-set(ngtcp2_crypto_openssl_INCLUDE_DIRS
+set(ngtcp2_crypto_quictls_INCLUDE_DIRS
   "${CMAKE_CURRENT_SOURCE_DIR}/../../lib/includes"
   "${CMAKE_CURRENT_BINARY_DIR}/../../lib/includes"
   "${CMAKE_CURRENT_SOURCE_DIR}/../../lib"
@@ -39,25 +39,25 @@ set(ngtcp2_crypto_openssl_INCLUDE_DIRS
   "${OPENSSL_INCLUDE_DIRS}"
 )
 
-foreach(name libngtcp2_crypto_openssl.pc)
+foreach(name libngtcp2_crypto_quictls.pc)
   configure_file("${name}.in" "${name}" @ONLY)
 endforeach()
 
 # Public shared library
 if(ENABLE_SHARED_LIB)
-  add_library(ngtcp2_crypto_openssl SHARED ${ngtcp2_crypto_openssl_SOURCES})
-  set_target_properties(ngtcp2_crypto_openssl PROPERTIES
+  add_library(ngtcp2_crypto_quictls SHARED ${ngtcp2_crypto_quictls_SOURCES})
+  set_target_properties(ngtcp2_crypto_quictls PROPERTIES
     COMPILE_FLAGS "${WARNCFLAGS}"
-    VERSION ${CRYPTO_OPENSSL_LT_VERSION}
-    SOVERSION ${CRYPTO_OPENSSL_LT_SOVERSION}
+    VERSION ${CRYPTO_QUICTLS_LT_VERSION}
+    SOVERSION ${CRYPTO_QUICTLS_LT_SOVERSION}
     C_VISIBILITY_PRESET hidden
     POSITION_INDEPENDENT_CODE ON
   )
-  target_include_directories(ngtcp2_crypto_openssl PUBLIC
-    ${ngtcp2_crypto_openssl_INCLUDE_DIRS})
-  target_link_libraries(ngtcp2_crypto_openssl ngtcp2 ${OPENSSL_LIBRARIES})
+  target_include_directories(ngtcp2_crypto_quictls PUBLIC
+    ${ngtcp2_crypto_quictls_INCLUDE_DIRS})
+  target_link_libraries(ngtcp2_crypto_quictls ngtcp2 ${OPENSSL_LIBRARIES})
 
-  install(TARGETS ngtcp2_crypto_openssl
+  install(TARGETS ngtcp2_crypto_quictls
     ARCHIVE DESTINATION "${CMAKE_INSTALL_LIBDIR}"
     LIBRARY DESTINATION "${CMAKE_INSTALL_LIBDIR}"
     RUNTIME DESTINATION "${CMAKE_INSTALL_BINDIR}")
@@ -65,22 +65,22 @@ endif()
 
 if(ENABLE_STATIC_LIB)
   # Public static library
-  add_library(ngtcp2_crypto_openssl_static ${ngtcp2_crypto_openssl_SOURCES})
-  set_target_properties(ngtcp2_crypto_openssl_static PROPERTIES
+  add_library(ngtcp2_crypto_quictls_static ${ngtcp2_crypto_quictls_SOURCES})
+  set_target_properties(ngtcp2_crypto_quictls_static PROPERTIES
     COMPILE_FLAGS "${WARNCFLAGS}"
-    VERSION ${CRYPTO_OPENSSL_LT_VERSION}
-    SOVERSION ${CRYPTO_OPENSSL_LT_SOVERSION}
-    ARCHIVE_OUTPUT_NAME ngtcp2_crypto_openssl${STATIC_LIB_SUFFIX}
+    VERSION ${CRYPTO_QUICTLS_LT_VERSION}
+    SOVERSION ${CRYPTO_QUICTLS_LT_SOVERSION}
+    ARCHIVE_OUTPUT_NAME ngtcp2_crypto_quictls${STATIC_LIB_SUFFIX}
     C_VISIBILITY_PRESET hidden
   )
-  target_compile_definitions(ngtcp2_crypto_openssl_static PUBLIC
+  target_compile_definitions(ngtcp2_crypto_quictls_static PUBLIC
     "-DNGTCP2_STATICLIB")
-  target_include_directories(ngtcp2_crypto_openssl_static PUBLIC
-    ${ngtcp2_crypto_openssl_INCLUDE_DIRS})
+  target_include_directories(ngtcp2_crypto_quictls_static PUBLIC
+    ${ngtcp2_crypto_quictls_INCLUDE_DIRS})
 
-  install(TARGETS ngtcp2_crypto_openssl_static
+  install(TARGETS ngtcp2_crypto_quictls_static
     DESTINATION "${CMAKE_INSTALL_LIBDIR}")
 endif()
 
-install(FILES "${CMAKE_CURRENT_BINARY_DIR}/libngtcp2_crypto_openssl.pc"
+install(FILES "${CMAKE_CURRENT_BINARY_DIR}/libngtcp2_crypto_quictls.pc"
   DESTINATION "${CMAKE_INSTALL_LIBDIR}/pkgconfig")
