+++ crypto/quictls/libngtcp2_crypto_quictls.pc.in
@@ -20,24 +20,14 @@
 # LIABLE FOR ANY CLAIM, DAMAGES OR OTHER LIABILITY, WHETHER IN AN ACTION
 # OF CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN CONNECTION
 # WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
-EXTRA_DIST = CMakeLists.txt
+prefix=@prefix@
+exec_prefix=@exec_prefix@
+libdir=@libdir@
+includedir=@includedir@
 
-AM_CFLAGS = $(WARNCFLAGS) $(DEBUGCFLAGS) $(EXTRACFLAG)
-AM_CPPFLAGS = -I$(top_srcdir)/lib/includes -I$(top_builddir)/lib/includes \
-	-I$(top_srcdir)/lib -DBUILDING_NGTCP2 \
-	-I$(top_srcdir)/crypto/includes -I$(top_builddir)/crypto/includes \
-	-I$(top_srcdir)/crypto -I$(top_builddir)/crypto \
-	@OPENSSL_CFLAGS@
-AM_LDFLAGS = ${LIBTOOL_LDFLAGS}
-
-pkgconfigdir = $(libdir)/pkgconfig
-pkgconfig_DATA = libngtcp2_crypto_openssl.pc
-DISTCLEANFILES = $(pkgconfig_DATA)
-
-lib_LTLIBRARIES = libngtcp2_crypto_openssl.la
-
-libngtcp2_crypto_openssl_la_SOURCES = openssl.c ../shared.c ../shared.h
-libngtcp2_crypto_openssl_la_LDFLAGS = -no-undefined \
-	-version-info $(CRYPTO_OPENSSL_LT_CURRENT):$(CRYPTO_OPENSSL_LT_REVISION):$(CRYPTO_OPENSSL_LT_AGE)
-libngtcp2_crypto_openssl_la_LIBADD = $(top_builddir)/lib/libngtcp2.la \
-	@OPENSSL_LIBS@
+Name: libngtcp2_crypto_quictls
+Description: ngtcp2 quictls crypto library
+URL: https://github.com/ngtcp2/ngtcp2
+Version: @VERSION@
+Libs: -L${libdir} -lngtcp2_crypto_quictls
+Cflags: -I${includedir}
