+++ examples/tls_client_context_quictls.h
@@ -22,8 +22,8 @@
  * OF CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN CONNECTION
  * WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
  */
-#ifndef TLS_CLIENT_CONTEXT_OPENSSL_H
-#define TLS_CLIENT_CONTEXT_OPENSSL_H
+#ifndef TLS_CLIENT_CONTEXT_QUICTLS_H
+#define TLS_CLIENT_CONTEXT_QUICTLS_H
 
 #ifdef HAVE_CONFIG_H
 #  include <config.h>
@@ -46,4 +46,4 @@ private:
   SSL_CTX *ssl_ctx_;
 };
 
-#endif // TLS_CLIENT_CONTEXT_OPENSSL_H
+#endif // TLS_CLIENT_CONTEXT_QUICTLS_H
